billing-app/requirements.txt
flask==2.3.0
flask-sqlalchemy==3.0.3
requests==2.31.0
psycopg2-binary==2.9.6
pika==1.3.1
python-dotenv==1.0.0
